Abstract the database
