+++ src/sssom_pydantic/web.py
@@ -9,7 +9,7 @@ from fastapi import APIRouter, Body, Depends, FastAPI, HTTPException, Path, Quer
 
 from sssom_pydantic import SemanticMapping
 from sssom_pydantic.api import SemanticMappingHash, mapping_hash_v1
-from sssom_pydantic.database import SemanticMappingDatabase
+from sssom_pydantic.database import SemanticMappingRepository
 from sssom_pydantic.examples import R1, R2
 from sssom_pydantic.process import MARKS, Mark
 
@@ -21,38 +21,38 @@ __all__ = [
 router = APIRouter()
 
 
-def get_controller(request: Request) -> SemanticMappingDatabase:
+def get_repository(request: Request) -> SemanticMappingRepository:
     """Get the controller from the web app."""
-    return cast(SemanticMappingDatabase, request.app.state.database)
+    return cast(SemanticMappingRepository, request.app.state.repository)
 
 
 #: A type alias for a controller that contains a dependency injection
 #: annotation for FastAPI.
-AnnotatedDatabase: TypeAlias = Annotated[SemanticMappingDatabase, Depends(get_controller)]
+AnnotatedRepository: TypeAlias = Annotated[SemanticMappingRepository, Depends(get_repository)]
 
 #: A type alias for a CURIE passed via the path
 AnnotatedCURIE = Annotated[str, Path(description="The CURIE for mapping record")]
 
 
 @router.get("/mapping/{curie}")
-def get_mapping(controller: AnnotatedDatabase, curie: AnnotatedCURIE) -> SemanticMapping:
+def get_mapping(repository: AnnotatedRepository, curie: AnnotatedCURIE) -> SemanticMapping:
     """Get a mapping by CURIE."""
-    mapping = controller.get_mapping(Reference.from_curie(curie))
+    mapping = repository.get_mapping(Reference.from_curie(curie))
     if mapping is None:
         raise HTTPException(status_code=404, detail="Mapping not found")
-    return mapping.to_semantic_mapping()
+    return mapping
 
 
 @router.delete("/mapping/{curie}")
-def delete_mapping(controller: AnnotatedDatabase, curie: AnnotatedCURIE) -> str:
+def delete_mapping(repository: AnnotatedRepository, curie: AnnotatedCURIE) -> str:
     """Get a mapping by CURIE."""
-    controller.delete_mapping(Reference.from_curie(curie))
+    repository.delete_mapping(Reference.from_curie(curie))
     return "ok"
 
 
 @router.post("/mapping/")
 def post_mapping(
-    database: AnnotatedDatabase,
+    repository: AnnotatedRepository,
     mapping: Annotated[
         SemanticMapping,
         Body(
@@ -69,40 +69,44 @@ def post_mapping(
     ],
 ) -> Reference:
     """Add a mapping by CURIE."""
-    return database.add_mapping(mapping)
+    return repository.add_mapping(mapping)
+
+
+# TODO bulk posting mappings (test on scale up to 10k or 100k at a time)
 
 
 @router.post("/action/publish/{curie}")
 def publish_mapping(
-    database: AnnotatedDatabase,
+    repository: AnnotatedRepository,
     curie: AnnotatedCURIE,
     date: Annotated[
         datetime.date | None, Query(..., description="The date on which the mapping was published")
     ] = None,
 ) -> Reference:
     """Publish a mapping with the given CURIE."""
-    return database.publish(Reference.from_curie(curie), date=date)
+    return repository.publish(Reference.from_curie(curie), date=date)
 
 
 @router.post("/action/curate/{curie}")
 def curate_mapping(
-    database: AnnotatedDatabase,
+    repository: AnnotatedRepository,
     curie: AnnotatedCURIE,
     authors: Annotated[list[Reference], Body(..., examples=[charlie])],
     mark: Annotated[Mark, Body(..., examples=list(MARKS))],
 ) -> Reference:
     """Publish a mapping with the given CURIE."""
-    return database.curate(Reference.from_curie(curie), authors=authors, mark=mark)
+    return repository.curate(Reference.from_curie(curie), authors=authors, mark=mark)
 
 
 def get_app(
     *,
-    database: SemanticMappingDatabase | None = None,
+    repository: SemanticMappingRepository | None = None,
     semantic_mapping_hash: SemanticMappingHash | None = None,
 ) -> FastAPI:
     """Get a FastAPI app.
 
-    :param database: A mapping database
+    :param repository: A mapping repository, e.g., a SQL database. If not given,
+        initializes an in-memory SQLite database.
     :param semantic_mapping_hash: A function that deterministically hashes a mapping.
         This is required until the SSSOM specification `defines a standard hashing
         procedure <https://github.com/mapping-commons/sssom/issues/436>`_.
@@ -116,12 +120,14 @@ def get_app(
         app = get_app()
         schema = app.openapi()
     """
-    if database is None:
-        database = SemanticMappingDatabase.memory(
+    if repository is None:
+        from sssom_pydantic.database import SemanticMappingDatabase
+
+        repository = SemanticMappingDatabase.memory(
             semantic_mapping_hash=semantic_mapping_hash or mapping_hash_v1
         )
     app = FastAPI()
-    app.state.database = database
+    app.state.repository = repository
     app.include_router(router)
     return app
